+++ include/envoy/upstream/cluster_manager.h
@@ -279,11 +279,10 @@ public:
   /**
    * Allocate a cluster from configuration proto.
    */
-  virtual ClusterSharedPtr clusterFromProto(const envoy::api::v2::Cluster& cluster,
-                                            ClusterManager& cm,
-                                            Outlier::EventLoggerSharedPtr outlier_event_logger,
-                                            AccessLog::AccessLogManager& log_manager,
-                                            bool added_via_api) PURE;
+  virtual ClusterSharedPtr
+  clusterFromProto(const envoy::api::v2::Cluster& cluster, ClusterManager& cm,
+                   Outlier::OutlierDetectionEventLoggerSharedPtr outlier_event_logger,
+                   AccessLog::AccessLogManager& log_manager, bool added_via_api) PURE;
 
   /**
    * Create a CDS API provider from configuration proto.